Rename autoArchive job task to archiver
